KIP-5 Изменения в моделях и логике
